requirements.txt
pygame
numpy
matplotlib
torch
torchvision
gymnasium
gymnasium_robotics
hydra-core
gymnasium[other]
